Make a dedicated "goto" command
We want two features:

- The original selection in the find-in-files view should not be
  touched.  This is done with the `restore_selection` context.

- We want the selection to be carried over to the target view.  That
  should work for cursors, and selections and here back- and forwards
  selections.

TODO: The final `set_sel` on the `view_` must actually run within a
`TextCommand` on its own -- a `TextCommand` on that view -- to get the
event listeners to run.  (For example, so the statusbar shows the right
column.)  The target view might be loading and then this all becomes
a noop; we thus need to await the loading.