Fix the two leading newlines when refreshing the first search
Sublime Text always draws two leading newlines when starting to draw a
search result and the find result view was already open.  This is the
case when we refresh a search.

We already implemented this when refreshing a nth search in the view
for n > 0.  For the first search, we really need an observer and a
manual fix.  We use the builtin `right_delete` here which fits our
need.