Fix: do not attempt to place find-in-files output panels
If the user chooses to use output panels for the find-in-files results,
we can't place them besides the current view.  (Actually Sublime Text
crashes when we call `set_view_index` on a panel!)

The quick fix is to patch `place_view` to not move un-movable views.
That is to say we prevent the side-effect.  It seems beneficial to
disable all the state tracking in that case as well but that's up for a
future commit.