Fix: ensure to use the same search when refreshing
It was always possible that the user changed search settings and
patterns before coming back to the "Find Results" view to refresh.

I.e. if a user "slurps" a new word into the search pattern buffer
(`ctrl+e`), then our refresh would have run with that new word.

Luckily, Sublime will print enough information in the search headline,
so that we can extract that information and forcefully send it to the
find panel.