Move all side-effects to `on_search_finished`
Without the commit, "refresh" will put the cursor on BOF position *if*
we run `fix_leading_newlines` as its byproduct -- which we do if there
is only one result shown in the find-in-files view.

This happens although `restore_previous_cursor` and
`fif_addon_set_cursor` run *after* `fix_leading_newlines`.

It turned out that running everything `on_search_finished` solves the
issue, looks good (t.i. the cursor does not flicker around) and
simplifies the code.

I could not find the actual culprit of the bug.