Ensure that local context line changes survive refreshes
Since the implementation was merely a POC one, de-obfuscate it.

It looks like the idea was always to have an aspect like wrapper; and
that concept translates to `with`-contextmanagers in Python.

Just wrapping the `find_all` command also narrows the application down
which should be a plus.

Note: Keep the old `ContextLineInjector` registered as Sublime Text will
not unload an `EventListener` during hot-reload.