Fix goto selection for not yet loaded views
Redoes 7fada66 (Fix closure error in `run_handlers`)

Before 7fada66 we did run all handlers using `set_timeout` (but with
the embarrassing closure bug).  Since I did not remember why I used
`set_timeout` in the first place, removing it seemed oppotune.

But it is not as Sublime sets the goto cursor one tick after loading
the view.  To override that and come back after we need to defer one
tick.  That is essentilly the same tick we added for result _panels_.

Reverting and fixing the closure bug using `partial` actually reduces
the code needed.  Surprisingly it does not add flicker when a result
_view_ is used.  Yay!