Register local symbols for filenames
The goal is to populate the "Goto Symbol" panel.  These symbols also
drive the "InlineOutline" plugin, and shortcuts I have that navigate
around symbols.

Now, we had `ctrl+r` bound here to refresh.  Revert that as it collides
with the default "Goto symbol" panel shortcut.  Use `F5` on all
platforms for now.

Although just registering the filenames is supportable by the native
Find Results syntax (it has a selector for the filenames), we go one
step further -- and I already regret that step -- and assign a variant
of that syntax to the buffers which also defines a selector for the
search header.  (Why?  Because it looks nice, like a separator, both
in the panel and any Outline view.)

Unfortunately, the native Find Results syntax is in the old format, so
we can't extend it but need a full copy here with one additional rule.

In the plugin I use `assign_syntax`.  If that call fails, "Plain Text"
gets applied, and the whole plugin doesn't attach anymore.