Implement `fif_addon_goto` from scratch
`fif_addon_goto` relied on the `result_line_regex` hack and internally
called `next_result`.

Remove that dependency:

a) so that `next_result`/`prev_result` works as before and expected,
b) to have finer control over the side-effects ordering, in fact this
   self-made version is snappier,
c) to support multiple selections once and later.