Add `fif_addon_goto_abort_search`
Add a command to "abort" the current search.  (Let's call it "abort" as
that's likely the right direction.)  This used to be just "close" but
since we now have previews, nesure that these close as well.

Still, I don't bind this but a user must on her own because "escape" is
difficult to do right for all users.