Handle double-click position in `fif_addon_goto`
If we set `want_event` we optionally receive the click coordinates if
the command has been triggered by a double-click.  With that information
we can compute the click point and proceed.

Note that we don't get called at all, if the user clicks on a direct
match or filename as that -- so it seems -- is handled by Sublime Text
directly.  (We would need to delete the output panel regexes to disable
the feature.)