Fix timing for highlighting the first match
Either I was wrong the first time or there is a timing change in Sublime
Text.

Anyway, the bug is:

	* select a term,
	* ctrl+shift+f for the quick search,
	* f5 to refresh
	* ctrl+w to go back to the original view
	* ctrl+shift+f for the same search

After that the cursor is at EOF position and scrolled to that.

This can be fixed by a) removing `fx` completely.  Then the standard
behavior kicks in and the cursor lands on the first line, in the search
header.  Or b), by always running `fx` on the next tick (`set_timeout`).

Choose b) for now.